+++ src/components/compare_commits.rs
@@ -256,7 +256,7 @@ impl CompareCommitsComponent {
 				if let Some(f) = self.details.files().selection_file()
 				{
 					let diff_params = DiffParams {
-						path: f.path.clone(),
+						path: f.full_path_str().to_string(),
 						diff_type: DiffType::Commits(ids),
 						options: DiffOptions::default(),
 					};
@@ -265,7 +265,11 @@ impl CompareCommitsComponent {
 						self.git_diff.last()?
 					{
 						if params == diff_params {
-							self.diff.update(f.path, false, last);
+							self.diff.update(
+								f.full_path_str().to_string(),
+								false,
+								last,
+							);
 							return Ok(());
 						}
 					}
@@ -293,7 +297,7 @@ impl CompareCommitsComponent {
 	}
 
 	fn can_focus_diff(&self) -> bool {
-		self.details.files().selection_file().is_some()
+		self.details.files().selection_file_ref().is_some()
 	}
 
 	fn hide_stacked(&mut self, stack: bool) {